Add compatibility with COP3014
format.py extracts a zipfile downloaded from canvas and reformats the
contents into a tarball as expected by the rest of the grading scripts.

* submitted files are renamed according to the roster
* timestamps are spoofed according to maintain late/on-time status
* generated tarball is produced to maintain compatibility

run.sh now check to see if format.py needs to be excuted first or not
depending on if a submission.zip file is present.

COP3363 is "Intro to C++ Programming with Unix" and the instructor
provides user submissions as a tarball to the TA's; which the
original scripts were intended for.

COP3014 is "Intro to C++ Programming without Unix" and the TA's
are expected to download the submissions via canvas, which
provides them as a zipfile with a vastly different naming
convention as well as no preserved timestamps.